# user-management-service/openapi.yaml
openapi: 3.0.3
info:
  title: User Management Service API
  description: |
    ユーザー管理サービスのAPI仕様書
    
    このAPIは認証機能のON/OFF状態を切り替え可能で、開発時は認証を無効化してテストすることができます。
    
    **検証済み環境:**
    - Docker環境での動作確認済み
    - 認証機能OFF状態での全エンドポイント検証完了
    - PostgreSQL 15 + Redis 7 構成
  version: 1.0.0
  contact:
    name: SkiShop Development Team
    email: dev@skishop.com
servers:
  - url: http://localhost:8081/api
    description: ローカル開発環境
  - url: https://api.skishop.com/user
    description: 本番環境
paths:
...
  /users:
    post:
      tags:
        - User Management
      summary: ユーザー作成
      description: 新しいユーザーアカウントを作成
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              username: "testuser1"
              email: "test1@example.com"
              password: "test123456"
              firstName: "Test1"
              lastName: "User1"
      responses:
        '201':
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 入力値不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: ユーザー名またはメールアドレスが既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
        - User Management
      summary: ユーザー情報取得
      description: 指定されたIDのユーザー情報を取得
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "e676091b-90c1-40b4-bd9f-ddeea80004ef"
      responses:
        '200':
          description: ユーザー情報取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - User Management
      summary: ユーザー情報更新
      description: 指定されたIDのユーザー情報を更新
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "e676091b-90c1-40b4-bd9f-ddeea80004ef"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              email: "updated1@example.com"
              firstName: "Updated1"
              lastName: "User1"
      responses:
        '200':
          description: ユーザー情報更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 入力値不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      tags:
        - User Management
      summary: ユーザー削除
      description: 指定されたIDのユーザーを削除
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "e676091b-90c1-40b4-bd9f-ddeea80004ef"
      responses:
        '204':
          description: ユーザー削除成功
        '404':
          description: ユーザーが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/check-email:
    get:
      tags:
        - User Management
      summary: メールアドレス存在確認
      description: 指定されたメールアドレスが既に登録されているか確認
      operationId: checkEmail
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          example: "test1@example.com"
      responses:
        '200':
          description: メールアドレス確認結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailCheckResponse'
              example:
                email: "test1@example.com"
                exists: false
                message: "Email does not exist"
  /users/verify-email:
    post:
      tags:
        - User Management
      summary: メール認証
      description: メール認証トークンを使用してメールアドレスを認証
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
            example:
              email: "test1@example.com"
              token: "sample-token"
      responses:
        '200':
          description: メール認証成功
        '400':
          description: 不正なトークンまたはメールアドレス
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
        - firstName
        - lastName
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 100
        password:
          type: string
          minLength: 6
          maxLength: 100
        firstName:
          type: string
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        phoneNumber:
          type: string
          maxLength: 20
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 100
        firstName:
          type: string
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        phoneNumber:
          type: string
          maxLength: 20
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
          nullable: true
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          nullable: true
        status:
          type: string
          enum: [PENDING, ACTIVE, INACTIVE, SUSPENDED]
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
          nullable: true
        roles:
          type: array
          items:
            type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
    EmailCheckResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        exists:
          type: boolean
        message:
          type: string
    EmailVerificationRequest:
      type: object
      required:
        - email
        - token
      properties:
        email:
          type: string
          format: email
        token:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT認証（認証機能ON時に使用）
        
        **認証機能OFF時の注意:**
        環境変数 `SKISHOP_AUTH_ENABLE=false` の場合、認証は無効化されます。
security:
  - BearerAuth: []
...